Abstracted the main loop into the application class, Created m_SceneData struct to simplify some other structures, this is where general data about the scene, ie matrices, positions etc will go.
Been awhile since last push, compared to last commit this is a pretty large refactoring, with better practices overall, fresh look.

Well the most recent one for checkpoint purposes, is there is no longer hardcoded references to the models in the Renderer header, now everything is stored in a models vector, and the ID of each model is actually a hardcoded index into that array.